Development (#34)
* init

* - added env file

* - removed loans menu and transfer it under Transactions
- added AddUpdateUserDrawer

* - bug fixes
- added role CRUD
- added permissions CRUD

* added roles and bug fixes

* - added user management
- bug fixes

* - fixed display of branches when editting a user

* - added groups management

* - fixed delete in all modules
- added client management
- bug fixes

* - added loans

* - bug fixes
- added roles for nav and data related
- added details based on the last meeting

* - bug fixes and enhancements

* - added loan applications
- bug fixes
- permissions
and many more...

* - added client detail page
- added modal
- bug fixes

* bug fixes

* updates

* updates

* bug fixes

* - updated the login screen

* - bug fixes

* Stopped tracking .env File

* Stopped tracking next.config.js File

* - added profile picture for clients

* - bug fixes

* - added daily cash collections
- bug fixes

* - fixed error in client list

* - bug fixes

* - bug fixes

* - bug fixes

* bug fixes

* - updates

* fixed issues

* - updates

* - updates

* updates

* updates

* bug fixed

* - updates

* - fixed issue

* - updates

* - updates

* - bug fix

* - bug fixes

* - bug fixes

* - bug fix

* - updates and bug fixes

* - bug fix

* - bug fixes

* - remarks update

* - updates

* - updates

* - fixes

* - debugging

* - removing debug

* - bug fix

* - bug fixed

* - update query by branch/group

+++ src/pages/api/branches/delete.js
@@ -19,12 +19,21 @@ async function deleteBranch(req, res) {
         .toArray();
 
     if (branches.length > 0) {
-        await db
-            .collection('branches')
-            .deleteOne({ _id: ObjectId(_id) });
+        const groups = await db.collection('groups').find({ branchId: branches[0]._id + '' }).toArray();
 
-        response = {
-            success: true
+        if (groups.length > 0) {
+            response = {
+                error: true,
+                message: `Can't delete branch ${branches[0].name} it is currently used in groups.`
+            };
+        } else {
+            await db
+                .collection('branches')
+                .deleteOne({ _id: ObjectId(_id) });
+
+            response = {
+                success: true
+            }
         }
     } else {
         response = {
